+++ src/App.jsx
@@ -2,11 +2,14 @@
 import { Routes, Route } from 'react-router-dom'
 
 // Import individual components and pages
-import Header from './components/Header/Header'
-import About from './pages/About'
-import Portfolio from './pages/Portfolio'
-import Contact from './pages/Contact'
-import Resume from './pages/Resume'
+import { Routes, Route } from "react-router-dom"
+// Update these paths
+import Header from "./components/Header/Header.jsx"  
+import About from "./pages/About/About.jsx"         
+import Portfolio from "./pages/Portfolio/Portfolio.jsx"
+import Contact from "./pages/Contact/Contact.jsx"
+import Resume from "./pages/Resume/Resume.jsx"
+import './App.css'
 
 // Import styles
 import './App.css'
